requirements.txt
scanpy>=1.9.0
anndata>=0.9.0
torch>=1.13.0
helical>=0.1.0
umap-learn>=0.5.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
statsmodels>=0.13.0
datasets>=2.0.0 